Track3_Interactive Argument Pair Recognition/requirements.txt
jieba==0.42.1
nltk==3.6.7
pandas==1.4.3
pytorch==1.12.0
rouge==1.0.1
transformers==4.16.2

